Update Micron config and add new stuff (#16)
* Create Micron docks and Anthead duct modifications

* Updated Micron config, added V2 and Stealthburner to user_mods and config

Updated my config for some cleanup and for the latest Klipper / Klipper-toolchanger versions.

Moved the main tool change macro to toolchanger_macros.cfg, now the T0 /T1 macros call _CHANGETOOL so updates are all in one place.

Tool change macros (T0/T1) can be called without an active tool on the carriage. So UNSELECT_TOOL  can be used and calling T0/T1 no longer causes and error.

V2 config works with Kalico and Klipper as of July 5, 2025.

Added all of my mods for running a Madmax on a Voron V2.